requirements.txt
streamlit==1.32.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
shap==0.45.0
matplotlib==3.8.4
plotly==5.22.0
pyyaml==6.0.1
streamlit-authenticator==0.2.3
